backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.16.5
    # via backend (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==25.1.0
    # via backend (pyproject.toml)
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
asyncpg==0.30.0
    # via backend (pyproject.toml)
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
cffi==2.0.0
    # via argon2-cffi-bindings
click==8.3.0
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.116.2
    # via backend (pyproject.toml)
fastapi-cli==0.0.13
    # via fastapi
fastapi-cloud-cli==0.2.0
    # via fastapi-cli
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.6
    # via fastapi
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
pycparser==2.23
    # via cffi
pydantic==2.11.9
    # via
    #   backend (pyproject.toml)
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.11.0
    # via backend (pyproject.toml)
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via backend (pyproject.toml)
python-dotenv==1.1.1
    # via
    #   backend (pyproject.toml)
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==14.1.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
sentry-sdk==2.38.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   backend (pyproject.toml)
    #   alembic
starlette==0.48.0
    # via fastapi
typer==0.19.2
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via sentry-sdk
uvicorn==0.35.0
    # via
    #   backend (pyproject.toml)
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn

# Kafka integration
aiokafka>=0.8.11

# HTTP client for Telegram API
aiohttp>=3.9.0
aiofiles>=23.2.0

# Additional async utilities
asyncio-mqtt>=0.16.1

# LangChain core
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-ollama>=0.2.0,<0.3.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0

# Vector store
langchain-qdrant>=0.2.0,<0.3.0
qdrant-client>=1.11.0,<2.0.0

# Agents & orchestration
langgraph>=0.2.28,<0.3.0

# Document loaders
pypdf>=4.0.0,<5.0.0
python-docx>=1.1.0,<2.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=5.0.0,<6.0.0

# Tools
duckduckgo-search>=6.0.0,<7.0.0

# Utilities
# tiktoken>=0.7.0,<0.8.0  # Requires Rust >= 1.71, disabled for now
python-magic>=0.4.27,<0.5.0

# ML stack (для будущих реализаций sklearn-пайплайнов)
scikit-learn>=1.3.0,<2.0.0
numpy>=1.26.0,<2.0.0
pandas>=2.2.0,<3.0.0
